Adding shared types so we can have definitions across packages
