+++ packages/escomplex-project/package.json
@@ -16,6 +16,7 @@
     "@swc/core": "^1.3.14",
     "eslint-config-prettier": "^8.5.0",
     "vitest": "^0.25.1",
+    "@ponticus/types": "0.0.1",
     "@ponticus/escomplex-test-data": "0.0.1",
     "@ponticus/babel-parser": "^0.0.1",
     "eslint": "^8.27.0",
