+++ packages/plugin-manager/package.json
@@ -4,7 +4,7 @@
   "description": "Provides a plugin manager that dispatches events to loaded plugins.",
   "homepage": "https://github.com/gabrielcsapo/ponticus",
   "main": "./dist/index.js",
-  "types": "dist/dts/index.d.ts",
+  "types": "./dist/dts/index.d.ts",
   "devDependencies": {
     "@swc/cli": "^0.1.57",
     "@swc/core": "^1.3.14",
