+++ packages/types/.eslintignore
@@ -1,0 +1,8 @@
+coverage/
+dist/
+cjs/
+docs/
+jspm_packages/
+node_modules/
+test/fixture/
+config.js
